+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Improve formatting of predicate directives, predicate definitions, and non-terminal definitions by inserting an empty line when switching to a different predicate/non-terminal
 * Fix formatting of block comments
+* Fix formatting of multi-line facts
 
 ## [0.56.0]